refactor(http): move streamhandler to server surface
Keep HTTP.streamhandler on the main server API instead of exposing
it through HTTP.Handlers, while leaving the router/middleware layer
focused on router-specific functionality.
